Ticketize - Update (Added the optional request-to-close)

+++ commands/ticket/close.js
@@ -1,3 +1,6 @@
+const DiscordJS = require("discord.js")
+const { MessageActionRow, MessageButton } = require("discord-buttons")
+
 module.exports = {
     name: "close",
     aliases: [],
@@ -5,18 +8,39 @@ module.exports = {
     async execute (message, args, prefix) {
         if (!message.channel.topic || !message.channel.topic.startsWith(`Tickets powered by ${client.user.username} | User ID:`)) return send_error(message, "This command is avaible only a ticket channel.")
 
-        tickets.findOne({ GuildId: message.guild.id, UserId: message.channel.topic.slice(40), ChannelId: message.channel.id }, async (err, data) => {
-            if (!data) return
-
-            await tickets.findOneAndDelete({ GuildId: message.guild.id, UserId: message.channel.topic.slice(40), ChannelId: message.channel.id })
+        configs.findOne({ GuildId: message.guild.id }, async (err, data) => {
+            if (data) {
+                if (data.Request_To_Close && data.Request_To_Close == true) {
+                    var button1 = new MessageButton()
+                    .setLabel("Close the Ticket")
+                    .setStyle("red")
+                    .setEmoji("🔒")
+                    .setID("close_confirm")
+                    var button2 = new MessageButton()
+                    .setLabel("Save in Transcript")
+                    .setStyle("gray")
+                    .setEmoji("📝")
+                    .setID("transcript")
+                    var row = new MessageActionRow()
+                    .addComponent(button1)
+                    .addComponent(button2)
+                    message.channel.send("❓ | **Are you sure to close this ticket?**", row)
+                } else {
+                    tickets.findOne({ GuildId: message.guild.id, UserId: message.channel.topic.slice(40), ChannelId: message.channel.id }, async (err, data) => {
+                        if (!data) return
+            
+                        await tickets.findOneAndDelete({ GuildId: message.guild.id, UserId: message.channel.topic.slice(40), ChannelId: message.channel.id })
+                    })
+            
+                    transcripts.findOne({ GuildId: message.guild.id, ChannelId: message.channel.id }, async (err, data) => {
+                        if (!data) return
+            
+                        await transcripts.findOneAndDelete({ GuildId: message.guild.id, ChannelId: message.channel.id })
+                    })
+                    
+                    message.channel.delete()
+                }
+            }
         })
-
-        transcripts.findOne({ GuildId: message.guild.id, ChannelId: message.channel.id }, async (err, data) => {
-            if (!data) return
-
-            await transcripts.findOneAndDelete({ GuildId: message.guild.id, ChannelId: message.channel.id })
-        })
-
-        message.channel.delete()
     }
 }
